reu 25/10

+++ blockmanager/types.h
@@ -1,3 +1,3 @@
 #include <math.h>
 #define BLOCK_SIZE 4096 // 4KB logical block size
-#define GET_BLOCKS(b) ceil((double)b / BLOCK_SIZE)
+#define GET_BLOCKS(b) (int)ceil((double)b / BLOCK_SIZE)
