+++ faloc/faloc.h
@@ -7,13 +7,13 @@ int fat_show(size_t *fat, int blocks);
 int fat_flag_block(size_t *fat, int block, int flag);
 
 // get block from file
-int fat_getf_block(int pen_fd, size_t *fat, int FE, unsigned int index, void *buf);
+int fat_getf_block(int pen_fd, size_t *fat, size_t FE, unsigned int index, void *buf);
 
 // write block to file
-int fat_writef_block(int pen_fd, size_t *fat, int FE, unsigned int index, void *buf);
+int fat_writef_block(int pen_fd, size_t *fat, size_t FE, unsigned int index, void *buf);
 
 // delete file from fat
-int fat_delete_file(int pen_fd, size_t *fat, int FE);
+int fat_delete_file(int pen_fd, size_t *fat, size_t FE);
 
 // increase filesize in fat
-int fat_increase_blocks(int pen_fd, size_t *fat, int blocksFat, int FE, int blocks);
+int fat_increase_blocks(int pen_fd, size_t *fat, int blocksFat, size_t FE, int blocks);
